Rust 0.11.0
305cf1f1

Rust documentation

A 30-minute Intro to Rust

The Rust tutorial (* PDF)

The Rust reference manual (* PDF)

1 Guides

Pointers

References and Lifetimes

Containers and Iterators

Tasks and Communication

Foreign Function Interface

Writing Safe Unsafe and Low-Level Code

Macros

Testing

Rust's Runtime

2 FAQs

Language Design FAQ

Language FAQ

Project FAQ

How to submit a bug report

3 Libraries

The standard library, std

The arena allocation library

The collections library

The flate compression library

The fourcc four-character code library

The getopts argument parsing library

The glob file path matching library

The green M:N runtime library

The hexfloat library for hexadecimal floating-point literals

The libc bindings

The native 1:1 threading runtime

The num arbitrary precision numerics library

The rand library for random numbers and distributions

The regex library for regular expressions

The rustc compiler

The rustuv M:N I/O library

The semver version collation library

The serialize value encoding/decoding library

The sync library for concurrency-enabled mechanisms and primitives

The syntax library, the Rust parser

The term terminal-handling library

The test library containing the unit-testing & micro-benchmark framework

The time library

The uuid 128-bit universally unique identifier library

The url library

The log library

4 Tooling

The rustdoc manual

5 External documentation

Note: While these are great resources for learning Rust, they may track a particular version of Rust that is likely not exactly the same as that for which this documentation was generated.

Rust for Rubyists - An excellent introduction for Rust; not just for Rubyists (tracks the most recent release).

Rust by Example - Short examples of common tasks in Rust (tracks the master branch).

The Rust wiki

6 Community

Reddit

Stack Overflow

The Rust IRC channels on irc.mozilla.org:

#rust - general discussion

#rust-gamedev - game development

#rust-internals - compiler and libraries

#rust-osdev - operating system development